feat: remove quote and unsuppress commands, replace with context menu
Quote command has not worked for a while due to API changes and unsuppress is better as a message context menu because it accepts a message anyway.
Also cleans up a bunch of old message retrieving code.

This also bumps JDA-Chewtils to 2.2.1 to fix a bug with context menus.

Note that, just as before, it can only unsuppress bot embeds, not embeds from links.